Int64 -> Int

+++ examples/demo_meshgroup.jl
@@ -11,8 +11,8 @@ if testCase == 1
     V[2 ] = GeometryBasics.Point{3, Float64}( 0.0,   -s, 0.0)
     V[3 ] = GeometryBasics.Point{3, Float64}(   s,  0.0, 0.0)
 
-    F = Vector{TriangleFace{Int64}}(undef,1)
-    F[1 ] = TriangleFace{Int64}(1,2,3)
+    F = Vector{TriangleFace{Int}}(undef,1)
+    F[1 ] = TriangleFace{Int}(1,2,3)
     # F,V=subtri(F,V,2)
 elseif testCase==2
     s=1.0
@@ -23,9 +23,9 @@ elseif testCase==2
     V[4 ] = GeometryBasics.Point{3, Float64}( 2*s,    s, 0.0)
     V[5 ] = GeometryBasics.Point{3, Float64}( 2*s,    -s, 0.0)
 
-    F = Vector{TriangleFace{Int64}}(undef,2)
-    F[1 ] = TriangleFace{Int64}(1,2,3)
-    F[2 ] = TriangleFace{Int64}(3,4,5)
+    F = Vector{TriangleFace{Int}}(undef,2)
+    F[1 ] = TriangleFace{Int}(1,2,3)
+    F[2 ] = TriangleFace{Int}(3,4,5)
     # F,V=subtri(F,V,2)
 elseif testCase==3
     r = 1.0
